// src/app/api/model/confirm-password-reset-request.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ConfirmPasswordResetRequest { 
    newPassword?: string;
    oobCode?: string;
}